Use bigint

+++ db/migrate/20250716125401_add_reporting_project_id_to_program.rb
@@ -8,6 +8,6 @@
 
 class AddReportingProjectIdToProgram < ActiveRecord::Migration[7.1]
   def change
-    add_column :sub_programs, :reporting_project_id, :integer
+    add_column :sub_programs, :reporting_project_id, :bigint
   end
 end
